requirements.txt
asgiref==3.5.2
Django==4.0
django-tastypie==0.14.4
python-dateutil==2.8.2
python-mimeparse==1.6.0
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
